move from CIAO tools to Python libraries distributed in CIAO
CIAO is distributed with some (not very well known) Python libraries
that make interacting with Chandra files and coordinates easier.
They access the same underlying implementation as the CIAO tools,
but are easier to handle and a more robust way to get information
out of Chandra files instead of calling CIAO tools like dmcoords,
which essentially start a run in a subshell.